Updated models to reflect database. Need to add Order model and update/add controllers.
