+++ NathanIanEcom/Models/Product.cs
@@ -8,23 +8,20 @@ namespace NathanIanEcom.Models
 {
     [DynamoDBTable("IanNathanProducts")]
 
-    public class Products
+    public class Product
     {
         [DynamoDBHashKey]
-
         public String ProductID { get; set; }
         [DynamoDBProperty]
-
-        public String Descirption { get; set; }
+        public String Description { get; set; }
         [DynamoDBProperty]
-
         public String ImageLink { get; set; }
         [DynamoDBProperty]
-
         public String Name { get; set; }
         [DynamoDBProperty]
-
         public String Price { get; set; }
+        [DynamoDBProperty]
+        public String Category { get; set; }
     }
 
     public class ProductModel
